Airports dialog: NULL reference exceptions with state/province.
Sometimes an airport doesn't have a state/province linked to it. In these cases, the airport.State value is NULL instead of an empty string. Since NULL (nonexistent) values can't use contains, starts with, and ends with operations, the search routine throws an exception. This update removes the possibility for a NULL value to make use of comparison operators. We will address the other fields as needed.