Mapping: Current location
Added the ability to get current location while flying through an area. For now, it includes only the city. Future revisions will include the state/territory and country, with the ability to change the "zoom level". The feature currently works with right bracket, shift+c instead of right bracket, c. When the feature is complete and tested, it will move to right bracket, c. Known issues: The airports database must be loaded, or you will receive a NullReference exception, and it doesn't detect large bodies of water such as oceans or seas.