Added Bing maps API key setting
Added the Bing maps API key setting. It is found in the general settings in TFM. If a Bing maps API key is not provided, the Where am I feature will not work while airborn. Instead, the pilot will receive a message stating that an API key is required to make the feature work.