FMC: partially copied to data grid format
To help people requiring a structured environment, we started porting a copy of the FMC over to a data grid layout. It provides many benefits such as the ability to have assistive technology identify the different portions of the FMC display. Example: What cell is the title, or what cell is lsk3r? The data grid also allows us to identify errors or problems with entered data. For instance, if the MTOW is almost obtained or exceeded, the cell turns yellow. This also happens in ground services where in progress items are colored in red. The data grid also provides a structured environment for screen reader users. It helps identify what lsk goes with what prompt.