FMC: HTML foundation
The previous FMC was difficult to use and understand, especially for new pilots. The rewrite uses HTML and WCAG 2.1 compliant code to display the FMC and its features. This entry disconnects the old FMC, then adds the basic foundation for HTML to take its place. Currently, the FMC does nothing, and the form that holds the HTML content is empty except for some invisible HTML 5 elements that make the FMC layout. At this point, the FMC is disconnected from the PMDG 737. When required, we will reconnect it. To access the FMC, load TFM while the simulator is running. Once TFM starts, press [ (left bracket), then Ctrl+C.