PMDG FMC: Section buttons
Converted the layout for the PMDG FMC into an HTML format. The previous method was confusing to understand, so we needed a better format. In this update, we added the HTML layout, accounted for the FMC's display, and added a toolbar at the bottom of the window containing the 18 buttons needed to access the FMC's content. None of the controls in the layout work yet. The ability to manipulate the FMC will come in a later build.