Build environment: Enabled TFM's debug mode
TFM now has a commandline parameter '/debug'. When required, we may ask users to use the /debug parameter when starting TFM to help troubleshoot problems. The first stage of implementation allows us to override the simulator must be loaded requirement. We can then work on TFM features without the need to have a simulator loaded in the background. The only need for a simulator is when we have to verify FSUIPC output/input.