New feature: spoiler percent setting
The spoilers on an aircraft help the pilot maintain speed on descent, approach, and landing. Most simulators only allow the pilot to retract, arm, or extend spoilers to 100%, nothing in between. This feature allows a pilot to press [, then Shift+Y to set the spoilers to a particular percentage. Valid values in the dialog are as follows:
0 = Spoilers retracted
1 = Spoilers armed
7 - 100 = The percentage to extend spoilers. EX: putting 10 in the dialog would result in roughly 10% extended.

The formulas used are precise to the 14th decimal place. Unfortunately, the system will round up or down depending on how far the fractional value is from the current whole number portion or the next available whole number portion. EX: 8.932009432164...% rounds to 9%, and 7.012095839997... rounds down to 7%.

Keyboard shortcuts: [, then Shift+y to change the value, and ], then Y to read it.