PMDG777: Offsets and panel objects
Added some classes that help facilitate reading offsets in the PMDG777. Also added a few enums to help keep track of a panel control's type. For example: The execute key on the CDU would have the PanelObjectType.Annunciator flag. The PanelObject, SingleStateToggle, and TExtDisplay allow us to customize the speech output for each panel object - annunciator, switch, button, multi-toggle, or whatever else developers put on a panel. The classes also keep track of which panel and panel section each panel control belongs. Now, we can filter, sort, and organize the panel controls. Ultimately, the new method of handling offset readouts will backport to the PMDG 747, PMDG 737, and the freeware aircraft.
The project isn't complete. We still need to finish adding flags to the PanelObjectType enum, add the offsets in the PMDG 777 SDK, and test the results.