PMDG 777 offsets
Added offsets for the PMDG 777. Includes the Overhead maint and Overhead panels. Created an easier way to add/remove/manage offsets for the PMDG aircraft. Included in the PMDG root folder are some classes that allow us to track and display offset values. These classes also make it easier to add new features in the future. Preperation is already made to add an aircraft verbosity feature, the ability to determine what type of panel control we are dealing with at the time, and the ability to organize the offsets into panels and panel sections. This new way of managing offsets will decrease the code in IoSubSystem from a large 300+ lines of code to a very small 5 lines of code. NOTE: some of the offsets appear to read wrong. They are reading as expected. Unfortunately, those who use First Officer Pro 2.0.3 will experience this strange behavior because First Officer Pro currently doesn't support the PMDG 777 completely.